Two colums full editor mode

+++ components/admin/Step.jsx
@@ -1,26 +1,14 @@
-import React, { useEffect, useCallback, useReducer } from 'react'
+import React, { useState, useEffect, useCallback, useReducer } from 'react'
 import { stepReducer } from 'lib/stepReducer'
 import cx from 'classnames'
 import StepEditor from '@/admin/StepEditor'
 import Button from '@/Button'
-import { BoltIcon, PlusIcon, TrashIcon, XCircleIcon } from '@heroicons/react/20/solid'
+import { BoltIcon, PlusIcon, TrashIcon, ViewColumnsIcon, XCircleIcon } from '@heroicons/react/24/outline'
 
-export default function Step({
-  penID,
-  num,
-  html,
-  css,
-  info,
-  brandcolor,
-  textcolor,
-  bgcolor,
-  total,
-  updateStep,
-  createStep,
-  deleteStep,
-}) {
+export default function Step({ penID, num, html, css, info, bgcolor, total, updateStep, createStep, deleteStep }) {
   const initialState = { html, css, _css: css, info, _info: info, focus: false, edited: false }
   const [state, accStep] = useReducer(stepReducer, initialState)
+  const [isVertical, setIsVertical] = useState(false)
 
   useEffect(() => {
     const handleKeydown = (e) => {
@@ -56,7 +44,10 @@ export default function Step({
   }
 
   const handleDelete = () => {
-    deleteStep({ penID, num })
+    // Confirm prompt
+    if (window.confirm(`Delete step ${num} from ${penID}?`)) {
+      deleteStep({ penID, num })
+    }
   }
 
   const handleReset = () => {
@@ -69,16 +60,28 @@ export default function Step({
 
   return (
     <div
-      className={cx('Step group', 'shrink-0 snap-center snap-mandatory sm:snap-proximity', 'rounded-lg border-4', {
-        'border-transparent': !state.focus && !state.edited,
-        'border-white/10': state.focus && !state.edited,
-        'border-yellow-600/50': state.edited && !state.focus,
-        'border-yellow-400/50': state.edited && state.focus,
-      })}
+      className={cx(
+        'Step group',
+        'w-screen ',
+        {
+          'lg:w-step': isVertical,
+          'lg:w-[80vw]': !isVertical,
+        },
+        'shrink-0 snap-center snap-mandatory sm:snap-proximity',
+        'rounded-lg border-4',
+        'sm:first:ml-[30%] sm:last:mr-[30%]',
+        {
+          'border-transparent': !state.focus && !state.edited,
+          'border-white/10': state.focus && !state.edited,
+          'border-yellow-600/50': state.edited && !state.focus,
+          'border-yellow-400/50': state.edited && state.focus,
+        }
+      )}
       style={{ backgroundColor: bgcolor }}>
-      <div className='w-screen sm:w-[400px] lg:w-[500px]'>
+      <>
         <div className='Buttons w-full h-12 p-2 flex gap-2 justify-between items-center rounded-t-xl bg-black/20'>
           <Button label={num} />
+
           <textarea
             rows='1'
             onInput={onInfo}
@@ -96,6 +99,11 @@ export default function Step({
 
           {!state.edited && (
             <>
+              <Button
+                icon={<ViewColumnsIcon />}
+                onClick={() => setIsVertical(!isVertical)}
+                className='sm:opacity-0 sm:group-hover:opacity-100 transition-opacity duration-300 ease-in-out'
+              />
               <Button
                 icon={<TrashIcon />}
                 onClick={handleDelete}
@@ -115,11 +123,12 @@ export default function Step({
           html={state.html}
           css={state.css}
           bgcolor={bgcolor}
+          isVertical={isVertical}
           onChange={onCss}
           onFocus={() => accStep({ type: 'FOCUS' })}
           onBlur={() => accStep({ type: 'BLUR' })}
         />
-      </div>
+      </>
     </div>
   )
 }
